requirements.txt
amqp==2.4.0
async-generator==1.10
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.5
django-allauth==0.38.0
django-rest-auth==0.9.3
djangorestframework==3.9.1
idna==2.8
kombu==4.2.2.post1
lxml==4.3.0
oauthlib==3.0.0
pyaes==1.6.1
pyasn1==0.4.5
PySocks==1.6.8
python3-openid==3.1.0
pytz==2018.9
redis==3.1.0
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
Telethon==1.5.5
urllib3==1.24.1
vine==1.2.0
